+++ src/main/java/faang/school/projectservice/model/stage_invitation/StageInvitation.java
@@ -1,4 +1,4 @@
-package faang.school.projectservice.model.stageinvitation;
+package faang.school.projectservice.model.stage_invitation;
 
 import faang.school.projectservice.model.stage.Stage;
 import faang.school.projectservice.model.TeamMember;
