+++ src/main/java/faang/school/projectservice/model/stage_invitation/StageInvitationStatus.java
@@ -1,4 +1,4 @@
-package faang.school.projectservice.model.stageinvitation;
+package faang.school.projectservice.model.stage_invitation;
 
 public enum StageInvitationStatus {
     PENDING,
