+++ src/main/java/faang/school/projectservice/validator/team_member/TeamMemberValidator.java
@@ -1,9 +1,9 @@
-package faang.school.projectservice.validator.teammember;
+package faang.school.projectservice.validator.team_member;
 
 import faang.school.projectservice.exception.DataValidationException;
 import faang.school.projectservice.model.Project;
 import faang.school.projectservice.model.TeamMember;
-import faang.school.projectservice.model.stageinvitation.StageInvitation;
+import faang.school.projectservice.model.stage_invitation.StageInvitation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
