+++ src/test/java/faang/school/projectservice/service/stage_invitation/StageInvitationServiceTest.java
@@ -1,24 +1,24 @@
-package faang.school.projectservice.service.stageinvitation;
+package faang.school.projectservice.service.stage_invitation;
 
-import faang.school.projectservice.dto.stageinvitation.StageInvitationDto;
-import faang.school.projectservice.dto.stageinvitation.StageInvitationFilterDto;
-import faang.school.projectservice.dto.stageinvitation.StageInvitationRejectDto;
+import faang.school.projectservice.dto.stage_invitation.StageInvitationDto;
+import faang.school.projectservice.dto.stage_invitation.StageInvitationFilterDto;
+import faang.school.projectservice.dto.stage_invitation.StageInvitationRejectDto;
 import faang.school.projectservice.exception.DataValidationException;
 import faang.school.projectservice.exception.EntityNotFoundException;
 import faang.school.projectservice.filter.Filter;
-import faang.school.projectservice.filter.stageinvitation.StageInvitationDescriptionFilter;
-import faang.school.projectservice.filter.stageinvitation.StageInvitationStageNameFilter;
+import faang.school.projectservice.filter.stage_invitation.StageInvitationDescriptionFilter;
+import faang.school.projectservice.filter.stage_invitation.StageInvitationStageNameFilter;
 import faang.school.projectservice.jpa.StageInvitationJpaRepository;
 import faang.school.projectservice.mapper.stageInvitation.StageInvitationMapper;
 import faang.school.projectservice.mapper.stageInvitation.StageInvitationMapperImpl;
 import faang.school.projectservice.model.TeamMember;
 import faang.school.projectservice.model.stage.Stage;
-import faang.school.projectservice.model.stageinvitation.StageInvitation;
-import faang.school.projectservice.model.stageinvitation.StageInvitationStatus;
+import faang.school.projectservice.model.stage_invitation.StageInvitation;
+import faang.school.projectservice.model.stage_invitation.StageInvitationStatus;
 import faang.school.projectservice.service.stage.StageService;
 import faang.school.projectservice.service.teammember.TeamMemberService;
-import faang.school.projectservice.validator.stageinvitation.StageInvitationValidator;
-import faang.school.projectservice.validator.teammember.TeamMemberValidator;
+import faang.school.projectservice.validator.stage_invitation.StageInvitationValidator;
+import faang.school.projectservice.validator.team_member.TeamMemberValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
